requirements.txt
# BMad Automation System - Python Dependencies
# Python 3.10+ required

# ============================================
# CORE DEPENDENCIES (Required)
# ============================================

# State Machine & Orchestration
langgraph>=0.0.26           # State machine for SM → PO → Dev → QA workflow
langchain-core>=0.1.0       # Required by langgraph

# Session Management
libtmux>=0.37.0             # Tmux session control for agent isolation

# File Monitoring
watchdog>=3.0.0             # Detect story file changes (agent completion)

# Configuration & Data
pyyaml>=6.0.1               # Parse configs, checkpoints, story files

# CLI & Output
rich>=13.7.0                # Terminal UI for progress display
click>=8.1.7                # CLI framework (future)

# ============================================
# DEVELOPMENT & TESTING
# ============================================

pytest>=7.4.3               # Testing framework
pytest-asyncio>=0.21.1      # Async test support
black>=23.12.0              # Code formatting
ruff>=0.1.8                 # Fast Python linter
mypy>=1.7.0                 # Type checking
ipython>=8.18.0             # Interactive debugging

# ============================================
# OPTIONAL (Future Phases)
# ============================================

# Phase 4: Enhanced error handling
# psutil>=5.9.6             # System resource monitoring

# Phase 5: Web dashboard
# fastapi>=0.104.1          # REST API
# uvicorn>=0.24.0           # ASGI server

# Phase 5: MCP servers
# httpx>=0.25.2             # HTTP client for MCP communication

# Distributed processing
# redis>=5.0.1              # State sharing across processes
# celery>=5.3.4             # Task queue